Reorganize the code directory
